backend-rdkit/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
rdkit-pypi==2022.9.5
pydantic==2.10.6
